Refactor generation pure helper utilities into dedicated module
